+++ agent/Entity.cpp
@@ -1,40 +1,21 @@
 #include <Windows.h>
 #include <winuser.h>
-#include <dxgi1_3.h>
-#include <d3d11_2.h>
-#include <d2d1_2.h>
-#include <d2d1_2helper.h>
-#include <dcomp.h>
 #include "./Brain.cpp"
-#include <wrl.h>
-using namespace Microsoft::WRL;
-#pragma comment(lib, "dxgi")
-#pragma comment(lib, "d3d11")
-#pragma comment(lib, "d2d1")
-#pragma comment(lib, "dcomp")
-#pragma comment(lib, "ole32")
+#include "./GifImage.cpp"
+#include "./Canvas.cpp"
+
 #pragma comment(lib, "user32.lib")
 
 class Entity: public WNDCLASSEX {
     public:
         typedef WNDCLASSEX super;
-        ComPtr<ID3D11Device> d3dDevice;
-        ComPtr<IDXGIDevice> dxgiDevice;
-        ComPtr<ID3D11DeviceContext> context;
-        ComPtr<IDXGIFactory2> dxFactory;
-        ComPtr<IDXGISwapChain1> swapChain;
-        ComPtr<ID2D1Factory2> d2Factory;
-        ComPtr<ID2D1Device1> d2Device;
-        ComPtr<ID2D1DeviceContext> d2dContext;
-        ComPtr<IDXGISurface2> surface;
-        ComPtr<ID2D1Bitmap1> bitmap;
-        ComPtr<IDCompositionDevice> dcompDevice;
-        ComPtr<IDCompositionTarget> target;
-        ComPtr<IDCompositionVisual> visual;
-        ComPtr<ID2D1SolidColorBrush> brush;
+        HWND hWindow;
 
         Brain* entityBrain;
-        HWND hWindow;
+        Canvas* windowCanvas; 
+
+        GifImage* rightBody;
+        GifImage* leftBody;
 
         int frame;
         UINT move = 0; 
@@ -50,18 +31,13 @@ class Entity: public WNDCLASSEX {
             this->lpszClassName = className ;
             this->hInstance = hInstance ;
             this->hIcon = LoadIcon (NULL, _T("../res/icon.ico"));
-            this->hIconSm = LoadIcon (NULL, _T("../res/icon.ico"));
-            this->hCursor = LoadCursor (NULL, IDC_ARROW);
-            this->lpszMenuName = NULL;                 
-            this->cbClsExtra = 0;                      
-            this->cbWndExtra = 0;                      
-            this->style = CS_DBLCLKS;                
+            this->hCursor = LoadCursor (NULL, IDC_ARROW);                  
+            this->style = CS_HREDRAW | CS_VREDRAW;                
             this->cbSize = sizeof (WNDCLASSEX);
-            this->hbrBackground = (HBRUSH) (COLOR_WINDOW + 1);
 
             this->frame = frame; 
-            this->rightBody = new Image(rightIdleBody) ;
-            this->leftBody = new Image(leftIdleBody) ;
+            this->rightBody = new GifImage(rightIdleBody) ;
+            this->leftBody = new GifImage(leftIdleBody) ;
         }
 
         bool materializeEntity() {
@@ -72,147 +48,13 @@ class Entity: public WNDCLASSEX {
                 COINIT_APARTMENTTHREADED | COINIT_DISABLE_OLE1DDE
             );
             if (hr != S_OK) return false;
-            
-            if (initializeCanvas() != true) return false;
+
+            windowCanvas = new Canvas(&this->hWindow) ;
+            windowCanvas->initializeCanvas() ;
+            this->rightBody->initilizeGif(); 
+            this->leftBody->initilizeGif();
 
             return true;   
-        }
-
-        bool initializeCanvas() {
-            UINT creationFlags = D3D11_CREATE_DEVICE_BGRA_SUPPORT;
-            D3D_FEATURE_LEVEL featureLevels[] =
-            {
-                D3D_FEATURE_LEVEL_11_1,
-                D3D_FEATURE_LEVEL_11_0,
-                D3D_FEATURE_LEVEL_10_1,
-                D3D_FEATURE_LEVEL_10_0,
-                D3D_FEATURE_LEVEL_9_3,
-                D3D_FEATURE_LEVEL_9_2,
-                D3D_FEATURE_LEVEL_9_1
-            };
-            
-            HRESULT hr = D3D11CreateDevice(
-                nullptr,
-                D3D_DRIVER_TYPE_HARDWARE,
-                nullptr,
-                creationFlags,
-                featureLevels,
-                ARRAYSIZE(featureLevels),
-                D3D11_SDK_VERSION,
-                &d3dDevice,
-                nullptr,
-                &context
-            );
-            if (hr != S_OK) return false ;
-
-            
-            hr = d3dDevice.As(&this->dxgiDevice) ;
-            if (hr != S_OK) return false ;
-
-            hr = CreateDXGIFactory2(
-                DXGI_CREATE_FACTORY_DEBUG,
-                __uuidof(dxFactory),
-                reinterpret_cast<void **>(dxFactory.GetAddressOf())
-            );
-            if (hr != S_OK) return false ;
-
-            //debuging
-            DXGI_SWAP_CHAIN_DESC1 description = {};
-            description.Format           = DXGI_FORMAT_B8G8R8A8_UNORM;     
-            description.BufferUsage      = DXGI_USAGE_RENDER_TARGET_OUTPUT;
-            description.SwapEffect       = DXGI_SWAP_EFFECT_FLIP_SEQUENTIAL;
-            description.BufferCount      = 2;                              
-            description.SampleDesc.Count = 1;                              
-            description.AlphaMode        = DXGI_ALPHA_MODE_PREMULTIPLIED;
-
-            RECT rect = {};
-            GetClientRect(this->hWindow, &rect);
-            description.Width  = rect.right - rect.left;  
-            description.Height = rect.bottom - rect.top;
-
-            hr = dxFactory->CreateSwapChainForComposition(
-                    dxgiDevice.Get(),
-                    &description,
-                    nullptr, 
-                    this->swapChain.GetAddressOf()
-                );
-            if (hr != S_OK) return false ;
-
-            // Create a single-threaded Direct2D factory with debugging information
-            D2D1_FACTORY_OPTIONS const options = { D2D1_DEBUG_LEVEL_INFORMATION };
-            hr = D2D1CreateFactory(
-                D2D1_FACTORY_TYPE_SINGLE_THREADED,
-                options,
-                d2Factory.GetAddressOf()
-            );
-            if (hr != S_OK) return false ;
-
-            // Create the Direct2D device that links back to the Direct3D device
-            hr = d2Factory->CreateDevice(
-                dxgiDevice.Get(),
-                d2Device.GetAddressOf()
-            );
-            if (hr != S_OK) return false ;
-
-            // Create the Direct2D device context that is the actual render target
-            // and exposes drawing commands
-            hr = d2Device->CreateDeviceContext(
-                D2D1_DEVICE_CONTEXT_OPTIONS_NONE,
-                this->d2dContext.GetAddressOf()
-            );
-            if (hr != S_OK) return false ;
-
-            // Retrieve the swap chain's back buffer
-            hr = this->swapChain->GetBuffer(
-                0,
-                __uuidof(surface),
-                reinterpret_cast<void **>(surface.GetAddressOf())
-            );
-            if (hr != S_OK) return false ;
-
-            // Create a Direct2D bitmap that points to the swap chain surface
-            D2D1_BITMAP_PROPERTIES1 properties = {};
-            properties.pixelFormat.alphaMode = D2D1_ALPHA_MODE_PREMULTIPLIED;
-            properties.pixelFormat.format    = DXGI_FORMAT_B8G8R8A8_UNORM;
-            properties.bitmapOptions         = D2D1_BITMAP_OPTIONS_TARGET |
-                                            D2D1_BITMAP_OPTIONS_CANNOT_DRAW;
-
-            hr = this->d2dContext->CreateBitmapFromDxgiSurface(
-                surface.Get(),
-                properties,
-                this->bitmap.GetAddressOf()
-            );
-            if (hr != S_OK) return false ;
-
-            // Point the device context to the bitmap for rendering
-            this->d2dContext->SetTarget(this->bitmap.Get());
-
-            hr = DCompositionCreateDevice(
-                this->dxgiDevice.Get(),
-                __uuidof(dcompDevice),
-                reinterpret_cast<void **>(dcompDevice.GetAddressOf())
-            );
-            if (hr != S_OK) return false;
-
-            hr = dcompDevice->CreateTargetForHwnd(
-                this->hWindow,
-                true, 
-                target.GetAddressOf()
-            );
-            if (hr != S_OK) return false;
-
-            hr = dcompDevice->CreateVisual(visual.GetAddressOf());
-            if (hr != S_OK) return false;
-
-            hr = visual->SetContent(swapChain.Get());
-            if (hr != S_OK) return false;
-
-            hr = target->SetRoot(visual.Get()) ;
-            if (hr != S_OK) return false;
-
-            hr = dcompDevice->Commit() ;
-
-            return (hr == S_OK)? true : false ;
         }
 
         void setHandle() {
@@ -222,16 +64,16 @@ class Entity: public WNDCLASSEX {
         }
 
         HWND createHandle() {
-            return CreateWindowEx(
+            return CreateWindowEx (
                 WS_EX_NOREDIRECTIONBITMAP, //expmn 0x00200000L
                 this->lpszClassName,
                 this->lpszClassName,
-                WS_OVERLAPPEDWINDOW | WS_VISIBLE,
-                CW_USEDEFAULT, CW_USEDEFAULT, CW_USEDEFAULT, CW_USEDEFAULT,//150, 150,
-                NULL,
-                NULL,
+                WS_VISIBLE,
+                300, 200, 200, 200,//150, 150,
+                nullptr,
+                nullptr,
                 this->hInstance, 
-                NULL
+                nullptr
             );
         }
 
@@ -241,31 +83,10 @@ class Entity: public WNDCLASSEX {
             return body ;
         }
 
-        bool animateEntity(HWND hwnd) {
-            this->d2dContext->BeginDraw();
-            this->d2dContext->Clear();
-            ComPtr<ID2D1SolidColorBrush> brush;
-            D2D1_COLOR_F const brushColor = D2D1::ColorF(0.18f,  // red
-                                                        0.55f,  // green
-                                                        0.34f,  // blue
-                                                        0.75f); // alpha
-            HRESULT hr = this->d2dContext->CreateSolidColorBrush(
-                brushColor,
-                brush.GetAddressOf()
-            );
-            if (hr != S_OK) return false;
-
-            D2D1_POINT_2F const ellipseCenter = D2D1::Point2F(150.0f, 
-                                                            150.0f); 
-            D2D1_ELLIPSE const ellipse = D2D1::Ellipse(ellipseCenter,
-                                                    100.0f,  
-                                                    100.0f ); 
-            this->d2dContext->FillEllipse(ellipse, brush.Get());
-            hr = this->d2dContext->EndDraw();
-            // Make the swap chain available to the composition engine
-            hr = this->swapChain->Present(1, 0); 
-
-            return (hr != S_OK)? true : false;
+        bool animateEntity(HWND hwnd) { 
+            windowCanvas->drawEntity() ;
+            //return (hr != S_OK)? true : false;
+            return true ;
         }
         
         void moveEntity(HWND hwnd) { 
@@ -276,6 +97,7 @@ class Entity: public WNDCLASSEX {
         }
 
         void disMaterializeEntity(){ 
+            
         }
 
         void setTimer(HWND hwnd, UINT timerId, UINT value) {
